// src/modules/shared/types/catalog.ts
export type Category = 'phones' | 'tablets' | 'accessories';

export type SortType = 'age' | 'title' | 'price';

export type PerPageOption = 4 | 8 | 16 | 'all';

export interface ProductSummary {
  id: number;
  category: Category;
  itemId: string;
  name: string;
  fullPrice: number;
  price: number;
  screen: string;
  capacity: string;
  color: string;
  ram: string;
  year: number;
  image: string;
}

export interface ProductDescriptionBlock {
  title: string;
  text: string[];
}

export interface ProductDetails {
  id: string;
  category: Category;
  namespaceId: string;
  name: string;
  capacityAvailable: string[];
  capacity: string;
  priceRegular: number;
  priceDiscount: number;
  colorsAvailable: string[];
  color: string;
  images: string[];
  description: ProductDescriptionBlock[];
  screen: string;
  resolution: string;
  processor: string;
  ram: string;
  camera?: string;
  zoom?: string;
  cell: string[];
}

export interface CartItem {
  id: string;
  quantity: number;
  product: ProductSummary;
}